+++ app/api/externalIntegrations.v2/automaticTranslation/contracts/AutomaticTranslationGateway.ts
