+++ app/api/externalIntegrations.v2/automaticTranslation/infrastructure/AutomaticTranslationService.ts
@@ -1,7 +1,7 @@
 import { LanguageISO6391 } from 'shared/types/commonTypes';
 import { AutomaticTranslationGateway } from '../contracts/AutomaticTranslationGateway';
 
-export class AutomaticTranslationService implements AutomaticTranslationGateway {
+export class ExternalAPIAutomaticTranslation implements AutomaticTranslationGateway {
   // eslint-disable-next-line class-methods-use-this
   async supportedLanguages() {
     const result: LanguageISO6391[] = ['ru', 'en', 'fr', 'es'];
